ML.H2O.gbm

Wrapper for an H2O gbm estimator. From their discription: Gradient Boosting Machine (for Regression and Classification) is a forward learning ensemble method. The guiding heuristic is that good predictive results can be obtained through increasingly refined approximations. H2O’s GBM sequentially builds regression trees on all the features of the dataset in a fully distributed way - each tree is built in parallel.

ML.H2O.gbm

Format

An object of class R6ClassGenerator of length 24.

Methods

initialize(ntrees=50, min_rows=9)

Creates a new gbm model @param nfolds integer (default = 0) specify the number of folds for cross-validation. @param ntrees integer (default = 50) specify the number of trees to build. @param min_rows (default = 9) specify the minimum number of observations for a leaf (nodesize in R). @param verbose (default = FALSE) the verbosity of the fitting procedure