Simulator.RunningExample

The running example from the paper simulation: if and with how much a binary treatment (antidepressant usage) attributes to changes in depression. We use a normalized hypothetical set of sum scores of the Inventory of DepressiveSymptomatology (IDS) questionnaire with a possible range of 0-100. Anti-depressant usage is a binary variable sampled from a binomial distribution. Variables: w1 gender 0/1 w2 quality of sleep 0/1 w3 level of activity (PAL value) between 1.4 and 2.4 A antidepressant = 1 no antdepressant = 0 Y IDS (Inventory of Depressive Symptomatology) Block time block Patient id patient identification Y_Y influence of Y on Y in the next block Y_A influence of Y on A in the next block Y_w2 influence of Y on w2 in the next blok

Simulator.RunningExample

Format

An object of class R6ClassGenerator of length 24.

Methods

new()

Starts a new simulator.

simulateWAY(numberOfBlocks, qw, ga, Qy, intervention, verbose)

Runs the simulation using the parameters provided.