predict.OnlineSuperLearner

S3 prediction function for the online superlearner package. Can be used to perform a prediction on the trained online superlearner object.

# S3 method for OnlineSuperLearner
predict(object, newdata, Y = NULL, ...)

Arguments

object
OnlineSuperLearner trained instance of an online superlearner class.

newdata
the new data to perform the prediction with. Note that this can be a data.frame, after which we will generate blocks based on the measurements in the data, or a Data.Base, which _should_ already include all necessary variables.

Y
the dependent variables for which we want to predict the outcome. The parameter is allowed to take several forms: - List of RandomVariable objects to predict - Single RandomVariable object to predict - List of strings with the names of the outputs (list('X','Y')) - Single string with the name of the output ('Y')

...
other parameters directly passed to the predict function

Value

list a list of estimator entries, each of which has a data.table of corresponding predicted probabilities.